+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add backward-compatible call from `DefaultEnpointResolution` to `DefaultEndpointResolution` on `MoyaProvider` class. `DefaultEndpointResolution` is now used internally as the default resolver. `DefaultEnpointResolution` can be removed in a future major release.
+
 # 2.1.0
 
 - Add option to pass an `Alamofire.Manager` to `MoyaProvider` initializer